+++ packages/suite/src/components/wallet/LayoutAccount/index.tsx
@@ -1,23 +1,50 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import { State } from '@suite-types/index';
+import { FormattedMessage } from 'react-intl';
 import TopNavigation from '@wallet-components/TopNavigation';
 import Layout from '@wallet-components/Layout';
+import { getRoute } from '@suite/utils/suite/router';
+import NETWORKS from '@suite-config/networks';
+import l10nMessages from './index.messages';
 
 interface Props {
     children: React.ReactNode;
 }
 
 const LayoutAccount = (props: Props) => (
-    <Layout>
-        <TopNavigation />
+    <Layout
+        topNavigationComponent={
+            <TopNavigation
+                items={[
+                    {
+                        route: getRoute('wallet-account-summary'),
+                        title: <FormattedMessage {...l10nMessages.TR_NAV_SUMMARY} />,
+                    },
+                    {
+                        route: getRoute('wallet-account-transactions'),
+                        title: <FormattedMessage {...l10nMessages.TR_NAV_TRANSACTIONS} />,
+                    },
+                    {
+                        route: getRoute('wallet-account-receive'),
+                        title: <FormattedMessage {...l10nMessages.TR_NAV_RECEIVE} />,
+                    },
+                    {
+                        route: getRoute('wallet-account-send'),
+                        title: <FormattedMessage {...l10nMessages.TR_NAV_SEND} />,
+                    },
+                    {
+                        route: getRoute('wallet-account-sign-verify'),
+                        title: <FormattedMessage {...l10nMessages.TR_NAV_SIGN_AND_VERIFY} />,
+                        isHidden: (coinShortcut: string) => {
+                            const network = NETWORKS.find(c => c.shortcut === coinShortcut);
+                            return network ? !network.hasSignVerify : false;
+                        },
+                    },
+                ]}
+            />
+        }
+    >
         {props.children}
     </Layout>
 );
 
-const mapStateToProps = (state: State) => ({
-    router: state.router,
-    suite: state.suite,
-});
-
-export default connect(mapStateToProps)(LayoutAccount);
+export default LayoutAccount;
